fix(svc-bridge): update MarketCapUsd type to long and change TradingPairsController method to POST
- Change MarketCapUsd property type from int? to long? in CoinMarketData and CoinBase classes for better handling of large values
- Update TradingPairsController to use POST instead of PUT for the trading pairs update endpoint, reflecting the intended operation
- Adjust corresponding integration tests to align with the updated HTTP method

+++ src/SVC_Microservices/SVC_Bridge/ApiContracts/Responses/Coins/CoinBase.cs
@@ -38,7 +38,7 @@ public abstract class CoinBase
     /// <summary>
     /// The market capitalization in USD.
     /// </summary>
-    public int? MarketCapUsd { get; init; }
+    public long? MarketCapUsd { get; init; }
 
     /// <summary>
     /// The current price in USD.
